+++ R/pcvjoyplot.R
@@ -119,7 +119,7 @@ pcv.joyplot <- function(df = NULL, index = NULL, group = NULL, y = NULL, id = NU
     facet_layer +
     ggridgeLayer +
     ggplot2::scale_x_continuous(n.breaks = 5, labels = ~ round(., 1)) +
-    ggplot2::labs(x = index, y = c(y, group)[1]) +
+    ggplot2::labs(x = index, y = gsub("dummy", "Density", c(y, group)[1])) +
     pcv_theme() +
     ggplot2::theme(legend.position = "none")
   return(p)